+++ templates4j-extension-strings/src/main/java/net/evilengineers/templates4j/extension/strings/ToLowercaseStringFunction.java
@@ -1,0 +1,14 @@
+package net.evilengineers.templates4j.extension.strings;
+
+import net.evilengineers.templates4j.spi.UserFunction;
+
+public class ToLowercaseStringFunction extends UserFunction {
+	@Override
+	public String getName() {
+		return "toLower";
+	}
+
+	public Object execute(String val) {
+		return val.toLowerCase();
+	}
+}