+++ CHANGES.md
@@ -4,7 +4,11 @@
   that doesn't result in you having to escape your actual content making it much less readable.  Now you can use delimiters such as <# and #> or longer, 
   reducing the need to ever use escape sequences.  (Henrik Gram)
 
-* Restructured project to maven default layout.  (Henrik Gram)
+* Restructured project to maven default layout and prepare for having multiple sub-projects.  (Henrik Gram)
+
+* Added sub-project for a maven-plugin that can automagically process your input files according to a specified grammar and template, creating java, 
+  html or any other type of text file you may need.  (Henrik Gram)
+
 
 ### 1.0.0 / 20014-04-18